ADLR/megatron-lm!3589 - Cherry-pick: `ADLR/megatron-lm!3221 - build: Upgrade to `25.05-py3-devel` image`

+++ tests/functional_tests/shell_test_utils/run_ci_test.sh
@@ -105,6 +105,11 @@ if [[ "$MODE" == "pretraining" && "$TEST_TYPE" != "release" ]]; then
     fi
 fi
 
+if [[ "$MODE" == "pretraining" && "$TEST_TYPE" = "release" ]]; then
+    TRAIN_ITERS=$(cat $TRAINING_PARAMS_PATH |
+        yq '.MODEL_ARGS."--exit-interval" // "100"')
+fi
+
 # Extract settings from params file
 NVTE_ALLOW_NONDETERMINISTIC_ALGO=$(cat $TRAINING_PARAMS_PATH |
     yq '.ENV_VARS.NVTE_ALLOW_NONDETERMINISTIC_ALGO')
@@ -124,6 +129,7 @@ for i in $(seq 1 $N_REPEAT); do
     export RUN_NUMBER=1
     export REPEAT=$i
     export CHECKPOINT_SAVE_PATH=$_CHECKPOINT_SAVE_PATH
+    export TRAINING_EXIT_CODE=0
 
     if [[ "$TEST_TYPE" = "frozen-start" ]]; then
         export CHECKPOINT_LOAD_PATH=$_CHECKPOINT_LOAD_PATH
@@ -136,7 +142,7 @@ for i in $(seq 1 $N_REPEAT); do
         export CHECKPOINT_SAVE_PATH=$_CHECKPOINT_SAVE_PATH
     fi
 
-    bash $ROOT_DIR/tests/functional_tests/shell_test_utils/_run_training.sh
+    bash $ROOT_DIR/tests/functional_tests/shell_test_utils/_run_training.sh || TRAINING_EXIT_CODE=$?
 
     if [[ "$TEST_TYPE" = "frozen-resume" && -z "$(ls -A "$_CHECKPOINT_LOAD_PATH" 2>/dev/null)" ]]; then
         echo "No frozen checkpoint found. Will skip second run."
@@ -147,24 +153,24 @@ for i in $(seq 1 $N_REPEAT); do
         break
     fi
 
-    if [[ "$TEST_TYPE" == "ckpt-resume" ]]; then
+    if [[ "$TEST_TYPE" == "ckpt-resume" && "$TRAINING_EXIT_CODE" -eq 0 ]]; then
         export CHECKPOINT_LOAD_PATH=$CHECKPOINT_SAVE_PATH
 
         rm -rf "$CHECKPOINT_LOAD_PATH/iter_$(printf "%07d\n" "$TRAIN_ITERS")"
         echo $((TRAIN_ITERS / 2)) >$CHECKPOINT_LOAD_PATH/latest_checkpointed_iteration.txt
 
         export RUN_NUMBER=2
-        bash $ROOT_DIR/tests/functional_tests/shell_test_utils/_run_training.sh
+        bash $ROOT_DIR/tests/functional_tests/shell_test_utils/_run_training.sh || TRAINING_EXIT_CODE=$?
     fi
 
-    if [[ "$TEST_TYPE" == "frozen-resume" ]]; then
+    if [[ "$TEST_TYPE" == "frozen-resume" && "$TRAINING_EXIT_CODE" -eq 0 ]]; then
 
         # Checkpoint-resume tests load from prev run
         export CHECKPOINT_LOAD_PATH=$_CHECKPOINT_LOAD_PATH
         export CHECKPOINT_SAVE_PATH=/tmp/checkpoints/
 
         export RUN_NUMBER=2
-        bash $ROOT_DIR/tests/functional_tests/shell_test_utils/_run_training.sh
+        bash $ROOT_DIR/tests/functional_tests/shell_test_utils/_run_training.sh || TRAINING_EXIT_CODE=$?
 
         export CHECKPOINT_SAVE_PATH=$_CHECKPOINT_SAVE_PATH
         rm -rf "$CHECKPOINT_SAVE_PATH/iter_0000$TRAIN_ITERS"
@@ -173,7 +179,6 @@ for i in $(seq 1 $N_REPEAT); do
 
     if [[ "$TEST_TYPE" == "release" ]]; then
         SKIP_PYTEST=1
-        TRAIN_ITERS=10000000
     fi
 
     if [[ ${RECORD_CHECKPOINTS} == "true" ]]; then
@@ -223,7 +228,7 @@ for i in $(seq 1 $N_REPEAT); do
     echo "Running pytest checks against golden values"
 
     # For pretraining jobs
-    if [[ "$MODE" == "pretraining" ]]; then
+    if [[ "$MODE" == "pretraining" && "$TRAINING_EXIT_CODE" -eq 0 ]]; then
         pytest -s -o log_cli=true --log-cli-level=info $ROOT_DIR/tests/functional_tests/python_test_utils/test_pretraining_regular_pipeline.py \
             --golden-values-path $GOLDEN_VALUES_PATH \
             --tensorboard-path $TENSORBOARD_PATH \
@@ -250,6 +255,12 @@ for i in $(seq 1 $N_REPEAT); do
                 --model-config-path ${TRAINING_PARAMS_PATH} \
                 $ALLOW_NONDETERMINISTIC_ALGO_ARG
         fi
+    fi
+
+    # Abort if training failed
+    if [[ "$TRAINING_EXIT_CODE" -ne 0 && "$TEST_TYPE" != "release" ]]; then
+        echo "Training failed. Aborting."
+        exit 1
     fi
 
 done
